ali_nlp/requirements.txt
aliyun-python-sdk-alinlp==1.0.16
python-dotenv==0.20.0
pydantic==1.9.1